Safari-609.1.11.1/LayoutTests/editing/pasteboard/data-transfer-get-data-on-paste-as-plain-text-when-custom-pasteboard-data-disabled-expected.txt

Tests that pasting as plain text only exposes "text/html" as well as "text/plain" in the clipboard when custom data support is turned off.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.clipboardData.types.includes("text/plain") is true
PASS event.clipboardData.types.includes("text/html") is true
PASS event.clipboardData.getData("text/plain") is "hello, world\nWebKit"
PASS event.clipboardData.getData("text/html") is not ""
PASS items = Array.from(event.clipboardData.items); items.length >= 2 is true
PASS items.some((item) => item.type == "text/plain") is true
PASS items.some((item) => item.type == "text/html") is true
PASS successfullyParsed is true

TEST COMPLETE
